backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
tensorflow==2.16.1
numpy>=1.24.3
opencv-python==4.8.1.78
pillow==10.1.0
python-dotenv==1.0.0 